requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 25.1.1
annotated-types=0.7.0=pypi_0
anyio=4.8.0=pypi_0
blinker=1.9.0=pypi_0
bzip2=1.0.8=h99b78c6_7
ca-certificates=2025.1.31=hf0a4a13_0
certifi=2025.1.31=pypi_0
chess=1.11.2=pypi_0
click=8.1.8=pypi_0
distro=1.9.0=pypi_0
flask=3.1.0=pypi_0
flask-cors=5.0.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.28.1=pypi_0
idna=3.10=pypi_0
itsdangerous=2.2.0=pypi_0
jinja2=3.1.6=pypi_0
jiter=0.9.0=pypi_0
libexpat=2.6.4=h286801f_0
libffi=3.4.2=h3422bc3_5
liblzma=5.6.4=h39f12f2_0
libsqlite=3.49.1=h3f77e49_1
libzlib=1.3.1=h8359307_2
markupsafe=3.0.2=pypi_0
ncurses=6.5=h5e97a16_3
openai=1.66.3=pypi_0
openssl=3.4.1=h81ee809_0
pip=25.0.1=pyh8b19718_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
python=3.11.11=hc22306f_2_cpython
python-dotenv=1.0.1=pypi_0
readline=8.2=h1d1bf99_2
setuptools=75.8.2=pyhff2d567_0
sniffio=1.3.1=pypi_0
stockfish=3.28.0=pypi_0
tk=8.6.13=h5083fa2_1
tqdm=4.67.1=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2025a=h78e105d_0
werkzeug=3.1.3=pypi_0
wheel=0.45.1=pyhd8ed1ab_1
